Move android directory to examples directory
