+++ examples/android/BUILD
@@ -2,7 +2,7 @@ package(default_visibility = ["//visibility:public"])
 
 android_library(
     name = "firestore_config_manager_impl",
-    srcs = ["src/main/java/com/google/startupos/android/FirestoreInitializer.java"],
+    srcs = ["src/main/java/com/google/startupos/examples/android/FirestoreInitializer.java"],
     deps = ["@com_google_firebase_firebase_common_11_8_0//aar"],
 )
 
@@ -37,14 +37,14 @@ android_library(
 android_binary(
     name = "android_app",
     srcs = glob([
-        "src/main/java/com/google/startupos/android/MainApplication.java",
-        "src/main/java/com/google/startupos/android/activities/*.java",
+        "src/main/java/com/google/startupos/examples/android/MainApplication.java",
+        "src/main/java/com/google/startupos/examples/android/activities/*.java",
     ]),
     assets = ["src/main/assets/google-services.json"],
     assets_dir = "src/main/assets/",
-    custom_package = "com.google.startupos.android",
-    manifest = "src/main/java/com/google/startupos/android/AndroidManifest.xml",
-    resource_files = glob(["src/main/java/com/google/startupos/android/res/**"]),
+    custom_package = "com.google.startupos.examples.android",
+    manifest = "src/main/java/com/google/startupos/examples/android/AndroidManifest.xml",
+    resource_files = glob(["src/main/java/com/google/startupos/examples/android/res/**"]),
     deps = [
         ":firestore",
         ":firestore_config_manager",
