+++ examples/android/README.md
