+++ examples/android/src/main/assets/google-services.json
