+++ examples/android/src/main/java/com/google/startupos/examples/android/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.google.startupos.android"
+    package="com.google.startupos.examples.android"
     android:versionCode="1"
     android:versionName="1.0">
 
@@ -16,7 +16,7 @@
         android:name=".MainApplication"
         android:label="@string/app_name" >
         <activity
-            android:name="com.google.startupos.android.activities.MainActivity"
+            android:name="com.google.startupos.examples.android.activities.MainActivity"
             android:label="@string/title_activity_main" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
