+++ examples/android/src/main/java/com/google/startupos/examples/android/res/drawable-hdpi/ic_launcher.png
