+++ examples/android/src/main/java/com/google/startupos/examples/android/res/drawable-ldpi/ic_launcher.png
