+++ examples/android/src/main/java/com/google/startupos/examples/android/res/drawable-mdpi/ic_launcher.png
