+++ examples/android/src/main/java/com/google/startupos/examples/android/res/drawable-xhdpi/ic_launcher.png
