+++ examples/android/src/main/java/com/google/startupos/examples/android/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.google.startupos.android.activities.MainActivity">
+    tools:context="com.google.startupos.examples.android.activities.MainActivity">
 
     <TextView android:text="@string/hello_world" android:layout_width="wrap_content"
         android:layout_height="wrap_content" android:id="@+id/text_view"/>
