+++ examples/android/src/main/java/com/google/startupos/examples/android/res/values-w820dp/dimens.xml
