+++ examples/android/src/main/java/com/google/startupos/examples/android/res/values/strings.xml
